Generate a HTML/PDF report exploring the basic results from derfinder

This function generates a HTML report exploring the basic results from single base-level approach derfinder analysis results (<www.bioconductor.org/packages/derfinder>). The HTML report itself is generated using rmarkdown (http://rmarkdown.rstudio.com/). It works best after using mergeResults.

derfinderReport(
  prefix,
  outdir = "basicExploration",
  output = "basicExploration",
  project = prefix,
  browse = interactive(),
  nBestRegions = 100,
  makeBestClusters = TRUE,
  nBestClusters = 2,
  fullCov = NULL,
  hg19 = TRUE,
  p.ideos = NULL,
  txdb = NULL,
  device = "png",
  significantVar = "qvalue",
  customCode = NULL,
  template = NULL,
  theme = NULL,
  digits = 2,
  ...
)

Arguments

prefix
The main data directory path where mergeResults was run. It should be the same as mergeResults(prefix).

outdir
The name of output directory relative to prefix.

output
The name of output HTML file (without the html extension).

project
The title of the project.

browse
If TRUE the HTML report is opened in your browser once it's completed.

nBestRegions
The number of region plots to make, ordered by area.

makeBestClusters
If TRUE, plotCluster is used on the nBestClusters regions by area. Note that these plots take some time to make.

nBestClusters
The number of region cluster plots to make by taking the nBestClusters regions ranked by area of the cluster.

fullCov
A list where each element is the result from loadCoverage used with cutoff=NULL. Can be generated using fullCoverage.

hg19
If TRUE then the reference is assumed to be hg19 and chromosome lengths as well as the default transcription database (TxDb.Hsapiens.UCSC.hg19.knownGene) will be used.

p.ideos
A list where each element is the result of plotIdeogram. If it's NULL and hg19=TRUE then they are created for the hg19 human reference.

txdb
Specify the transcription database to use for making the plots for the top regions by area. If NULL and hg19=TRUE then TxDb.Hsapiens.UCSC.hg19.knownGene is used.

device
The graphical device used when knitting. See more at http://yihui.name/knitr/options (dev argument).

significantVar
A character variable specifying whether to use the p-values, the FDR adjusted p-values or the FWER adjusted p-values to determine significance. Has to be either 'pvalue', 'qvalue' or 'fwer'.

customCode
An absolute path to a child R Markdown file with code to be evaluated before the reproducibility section. Its useful for users who want to customize the report by adding conclusions derived from the data and/or further quality checks and plots.

template
Template file to use for the report. If not provided, will use the default file found in basicExploration/basicExploration.Rmd within the package source.

theme
A ggplot2 theme to use for the plots made with ggplot2.

digits
The number of digits to round to in the interactive table of the top nBestRegions. Note that p-values and adjusted p-values won't be rounded.

...
Arguments passed to other methods and/or advanced arguments. Advanced arguments:

chrsStyle

The naming style of the chromosomes. By default, UCSC. See seqlevelsStyle.

species

Species name. See extendedMapSeqlevels for more information.

currentStyle

Current naming style used. See extendedMapSeqlevels for more information.

fullRegions

Part of the output of mergeResults. Specify it only if you have already loaded it in memory.

fullNullSummary

Part of the output of mergeResults. Specify it only if you have already loaded it in memory.

fullAnnotatedRegions

Part of the output of mergeResults. Specify it only if you have already loaded it in memory.

optionsStats

Part of the output of analyzeChr. Specify it only if you have already loaded it in memory.

optionsMerge

Part of the output of mergeResults. Specify it only if you have already loaded it in memory.

overviewParams

A two element list with base_size and areaRel that control the text size for the genomic overview plots.

output_format

Either html_document, pdf_document or knitrBootstrap::bootstrap_document unless you modify the YAML template.

clean

Logical, whether to clean the results or not. Passed to render.

Passed to extendedMapSeqlevels.

Value

An HTML report with a basic exploration of the derfinder results. See the example output at http://leekgroup.github.io/regionReport/reference/derfinderReport-example/basicExploration/basicExploration.html.

Details

Set output_format to 'knitrBootstrap::bootstrap_document' or 'pdf_document' if you want a HTML report styled by knitrBootstrap or a PDF report respectively. If using knitrBootstrap, we recommend the version available only via GitHub at https://github.com/jimhester/knitrBootstrap which has nicer features than the current version available via CRAN. You can also set the output_format to 'html_document' for a HTML report styled by rmarkdown. The default is set to 'BiocStyle::html_document'.

If you modify the YAML front matter of template, you can use other values for output_format.

The HTML report styled with knitrBootstrap can be smaller in size than the 'html_document' report.

See also

Author

Leonardo Collado-Torres

Examples

## Load derfinder
library("derfinder")

## The output will be saved in the 'derfinderReport-example' directory
dir.create("derfinderReport-example", showWarnings = FALSE, recursive = TRUE)

## For convenience, the derfinder output has been pre-computed
file.copy(system.file(file.path("extdata", "chr21"),
    package = "derfinder",
    mustWork = TRUE
), "derfinderReport-example", recursive = TRUE)
#> [1] TRUE
if (FALSE) {
## If you prefer, you can generate the output from derfinder
initialPath <- getwd()
setwd(file.path(initialPath, "derfinderReport-example"))

## Collapse the coverage information
collapsedFull <- collapseFullCoverage(list(genomeData$coverage),
    verbose = TRUE
)

## Calculate library size adjustments
sampleDepths <- sampleDepth(collapsedFull,
    probs = c(0.5), nonzero = TRUE,
    verbose = TRUE
)

## Build the models
group <- genomeInfo$pop
adjustvars <- data.frame(genomeInfo$gender)
models <- makeModels(sampleDepths, testvars = group, adjustvars = adjustvars)

## Analyze chromosome 21
analyzeChr(
    chr = "21", coverageInfo = genomeData, models = models,
    cutoffFstat = 1, cutoffType = "manual", seeds = 20140330, groupInfo = group,
    mc.cores = 1, writeOutput = TRUE, returnOutput = FALSE
)

## Change the directory back to the original one
setwd(initialPath)
}

## Merge the results from the different chromosomes. In this case, there's
## only one: chr21
mergeResults(
    chrs = "21", prefix = "derfinderReport-example",
    genomicState = genomicState$fullGenome
)
#> extendedMapSeqlevels: sequence names mapped from NCBI to UCSC for species homo_sapiens
#> 2021-08-10 00:37:59 mergeResults: Saving options used
#> 2021-08-10 00:37:59 Loading chromosome chr21
#> Neither 'cutoffFstatUsed' nor 'optionsStats' were supplied, so the FWER calculation step will be skipped.
#> 2021-08-10 00:37:59 mergeResults: Saving fullNullSummary
#> 2021-08-10 00:37:59 mergeResults: Re-calculating the p-values
#> 2021-08-10 00:37:59 mergeResults: Saving fullRegions
#> 2021-08-10 00:37:59 mergeResults: assigning genomic states
#> 2021-08-10 00:37:59 annotateRegions: counting
#> 2021-08-10 00:37:59 annotateRegions: annotating
#> 2021-08-10 00:37:59 mergeResults: Saving fullAnnotatedRegions
#> 2021-08-10 00:37:59 mergeResults: Saving fullFstats
#> 2021-08-10 00:37:59 mergeResults: Saving fullTime

## Load the options used for calculating the statistics
load(file.path("derfinderReport-example", "chr21", "optionsStats.Rdata"))

## Generate the HTML report
report <- derfinderReport(
    prefix = "derfinderReport-example", browse = FALSE,
    nBestRegions = 15, makeBestClusters = TRUE,
    fullCov = list("21" = genomeDataRaw$coverage), optionsStats = optionsStats
)
#> Registered S3 method overwritten by 'GGally':
#>   method from   
#>   +.gg   ggplot2
#> Writing 9 Bibtex entries ... 
#> OK
#> Results written to file 'derfinderReport-example/basicExploration/basicExploration.bib'
#> extendedMapSeqlevels: sequence names mapped from NCBI to UCSC for species homo_sapiens
#> 
#> 
#> processing file: basicExploration.Rmd
#>   |                                                                                                                      |                                                                                                              |   0%  |                                                                                                                      |.                                                                                                             |   1%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |...                                                                                                           |   3%
#> label: docSetup (with options) 
#> List of 3
#>  $ bootstrap.show.code   : logi FALSE
#>  $ dev                   : chr "png"
#>  $ bootstrap.show.message: logi FALSE
#> 
#>   |                                                                                                                      |....                                                                                                          |   4%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |......                                                                                                        |   5%
#> label: setup (with options) 
#> List of 1
#>  $ bootstrap.show.message: logi FALSE
#> 
#>   |                                                                                                                      |.......                                                                                                       |   6%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |.........                                                                                                     |   8%
#> label: pvals (with options) 
#> List of 4
#>  $ fig.width : num 10
#>  $ fig.height: num 10
#>  $ eval      : logi TRUE
#>  $ echo      : logi TRUE
#> 
#>   |                                                                                                                      |..........                                                                                                    |   9%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |...........                                                                                                   |  10%
#> label: summPval (with options) 
#> List of 2
#>  $ eval: logi TRUE
#>  $ echo: logi TRUE
#> 
#>   |                                                                                                                      |.............                                                                                                 |  12%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |..............                                                                                                |  13%
#> label: summQval (with options) 
#> List of 2
#>  $ eval: logi TRUE
#>  $ echo: logi TRUE
#> 
#>   |                                                                                                                      |................                                                                                              |  14%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |.................                                                                                             |  16%
#> label: qsummary (with options) 
#> List of 3
#>  $ eval   : logi TRUE
#>  $ echo   : logi TRUE
#>  $ results: chr "asis"
#> 
#>   |                                                                                                                      |...................                                                                                           |  17%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |....................                                                                                          |  18%
#> label: summFwer (with options) 
#> List of 2
#>  $ eval: logi FALSE
#>  $ echo: logi FALSE
#> 
#>   |                                                                                                                      |.....................                                                                                         |  19%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |.......................                                                                                       |  21%
#> label: FWERsummary (with options) 
#> List of 3
#>  $ eval   : logi FALSE
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                                                      |........................                                                                                      |  22%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |..........................                                                                                    |  23%
#> label: regLen (with options) 
#> List of 4
#>  $ fig.width : num 14
#>  $ fig.height: num 14
#>  $ eval      : logi TRUE
#>  $ echo      : logi TRUE
#> 
#>   |                                                                                                                      |...........................                                                                                   |  25%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |.............................                                                                                 |  26%
#> label: regLen2 (with options) 
#> List of 4
#>  $ fig.width : num 10
#>  $ fig.height: num 10
#>  $ eval      : logi FALSE
#>  $ echo      : logi FALSE
#> 
#>   |                                                                                                                      |..............................                                                                                |  27%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |...............................                                                                               |  29%
#> label: regArea (with options) 
#> List of 4
#>  $ fig.width : num 14
#>  $ fig.height: num 14
#>  $ eval      : logi TRUE
#>  $ echo      : logi TRUE
#> 
#>   |                                                                                                                      |.................................                                                                             |  30%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |..................................                                                                            |  31%
#> label: regArea2 (with options) 
#> List of 4
#>  $ fig.width : num 10
#>  $ fig.height: num 10
#>  $ eval      : logi FALSE
#>  $ echo      : logi FALSE
#> 
#>   |                                                                                                                      |....................................                                                                          |  32%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |.....................................                                                                         |  34%
#> label: nullLengthArea (with options) 
#> List of 4
#>  $ fig.width : num 14
#>  $ fig.height: num 14
#>  $ eval      : logi TRUE
#>  $ echo      : logi TRUE
#> 
#>   |                                                                                                                      |.......................................                                                                       |  35%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |........................................                                                                      |  36%
#> label: meanCov (with options) 
#> List of 4
#>  $ fig.width : num 14
#>  $ fig.height: num 14
#>  $ eval      : logi TRUE
#>  $ echo      : logi TRUE
#> 
#>   |                                                                                                                      |.........................................                                                                     |  38%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |...........................................                                                                   |  39%
#> label: meanCov2 (with options) 
#> List of 4
#>  $ fig.width : num 10
#>  $ fig.height: num 10
#>  $ eval      : logi FALSE
#>  $ echo      : logi FALSE
#> 
#>   |                                                                                                                      |............................................                                                                  |  40%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |..............................................                                                                |  42%
#> label: MAstyle (with options) 
#> List of 2
#>  $ fig.width : num 10
#>  $ fig.height: num 10
#> 
#>   |                                                                                                                      |...............................................                                                               |  43%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |.................................................                                                             |  44%
#> label: genomeOverview1 (with options) 
#> List of 6
#>  $ message   : logi FALSE
#>  $ fig.width : num 7
#>  $ fig.height: num 9
#>  $ dpi       : num 300
#>  $ eval      : logi TRUE
#>  $ echo      : logi TRUE
#> 
#> Scale for 'x' is already present. Adding another scale for 'x', which will replace the existing scale.
#> Scale for 'x' is already present. Adding another scale for 'x', which will replace the existing scale.
#>   |                                                                                                                      |..................................................                                                            |  45%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |...................................................                                                           |  47%
#> label: manhattanPlots (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                                                      |.....................................................                                                         |  48%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |......................................................                                                        |  49%
#> label: genomeOverview2 (with options) 
#> List of 4
#>  $ message   : logi FALSE
#>  $ fig.width : num 7
#>  $ fig.height: num 9
#>  $ dpi       : num 300
#> 
#> Scale for 'x' is already present. Adding another scale for 'x', which will replace the existing scale.
#>   |                                                                                                                      |........................................................                                                      |  51%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |.........................................................                                                     |  52%
#> label: annoReg (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                                                                      |...........................................................                                                   |  53%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |............................................................                                                  |  55%
#> label: genomeOverview3 (with options) 
#> List of 6
#>  $ message   : logi FALSE
#>  $ fig.width : num 7
#>  $ fig.height: num 9
#>  $ dpi       : num 300
#>  $ eval      : logi TRUE
#>  $ echo      : logi TRUE
#> 
#> Scale for 'x' is already present. Adding another scale for 'x', which will replace the existing scale.
#>   |                                                                                                                      |.............................................................                                                 |  56%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |...............................................................                                               |  57%
#> label: plotRegions (with options) 
#> List of 3
#>  $ fig.width             : num 10
#>  $ fig.height            : num 10
#>  $ bootstrap.show.message: logi FALSE
#> 
#>   |                                                                                                                      |................................................................                                              |  58%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |..................................................................                                            |  60%
#> label: countTable (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                                                                      |...................................................................                                           |  61%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |.....................................................................                                         |  62%
#> label: vennDiagram (with options) 
#> List of 2
#>  $ fig.width : num 7
#>  $ fig.height: num 7
#> 
#>   |                                                                                                                      |......................................................................                                        |  64%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |.......................................................................                                       |  65%
#> label: vennDiagramSignificant (with options) 
#> List of 4
#>  $ eval      : logi TRUE
#>  $ echo      : logi TRUE
#>  $ fig.width : num 7
#>  $ fig.height: num 7
#> 
#>   |                                                                                                                      |.........................................................................                                     |  66%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |..........................................................................                                    |  68%
#> label: bestAreaInfo (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                                                                      |............................................................................                                  |  69%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |.............................................................................                                 |  70%
#> label: plotCluster (with options) 
#> List of 4
#>  $ message: logi FALSE
#>  $ warning: logi FALSE
#>  $ eval   : logi TRUE
#>  $ echo   : logi TRUE
#> 
#>   |                                                                                                                      |...............................................................................                               |  71%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |................................................................................                              |  73%
#> label: bestClusters (with options) 
#> List of 6
#>  $ message   : logi FALSE
#>  $ warning   : logi FALSE
#>  $ fig.width : num 20
#>  $ fig.height: num 10
#>  $ eval      : logi TRUE
#>  $ echo      : logi TRUE
#> 
#> Warning: `guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
#> Parsing transcripts...
#> Parsing exons...
#> Parsing cds...
#> Parsing utrs...
#> ------exons...
#> ------cdss...
#> ------introns...
#> ------utr...
#> aggregating...
#> Done
#> Constructing graphics...
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
#> Warning: Transformation introduced infinite values in continuous y-axis
#> Warning: Transformation introduced infinite values in continuous y-axis
#> Warning: `guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead.
#> Parsing transcripts...
#> Parsing exons...
#> Parsing cds...
#> Parsing utrs...
#> ------exons...
#> ------cdss...
#> ------introns...
#> ------utr...
#> aggregating...
#> Done
#> Constructing graphics...
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will replace the existing one.
#> Warning: Transformation introduced infinite values in continuous y-axis
#> Warning: Transformation introduced infinite values in continuous y-axis
#>   |                                                                                                                      |.................................................................................                             |  74%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |...................................................................................                           |  75%
#> label: permute2 (with options) 
#> List of 2
#>  $ eval: logi TRUE
#>  $ echo: logi TRUE
#> 
#>   |                                                                                                                      |....................................................................................                          |  77%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |......................................................................................                        |  78%
#> label: permuteTable (with options) 
#> List of 3
#>  $ results: chr "asis"
#>  $ eval   : logi TRUE
#>  $ echo   : logi TRUE
#> 
#>   |                                                                                                                      |.......................................................................................                       |  79%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |.........................................................................................                     |  81%
#> label: unnamed-chunk-1 (with options) 
#> List of 2
#>  $ child: NULL
#>  $ eval : logi FALSE
#> 
#>   |                                                                                                                      |..........................................................................................                    |  82%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |...........................................................................................                   |  83%
#> label: analysisCall (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                                                      |.............................................................................................                 |  84%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |..............................................................................................                |  86%
#> label: mergeCall (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                                                      |................................................................................................              |  87%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |.................................................................................................             |  88%
#> label: thecall (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                                                      |...................................................................................................           |  90%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |....................................................................................................          |  91%
#> label: reproducibility1 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                                                      |.....................................................................................................         |  92%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |.......................................................................................................       |  94%
#> label: reproducibility2 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                                                      |........................................................................................................      |  95%
#>   ordinary text without R code
#> 
#>   |                                                                                                                      |..........................................................................................................    |  96%
#> label: reproducibility3 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                                                      |...........................................................................................................   |  97%
#>    inline R code fragments
#> 
#>   |                                                                                                                      |............................................................................................................. |  99%
#> label: bibliography (with options) 
#> List of 3
#>  $ results: chr "asis"
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#> 
#>   |                                                                                                                      |..............................................................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#> output file: basicExploration.knit.md
#> /usr/local/bin/pandoc +RTS -K512m -RTS basicExploration.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output basicExploration.html --lua-filter /__w/_temp/Library/bookdown/rmarkdown/lua/custom-environment.lua --lua-filter /__w/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /__w/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --self-contained --wrap preserve --variable bs3=TRUE --standalone --section-divs --table-of-contents --toc-depth 3 --variable toc_float=1 --variable toc_selectors=h1,h2,h3 --variable toc_collapsed=1 --variable toc_smooth_scroll=1 --variable toc_print=1 --template /tmp/Rtmpnpd80O/BiocStyle/template.html --no-highlight --variable highlightjs=1 --number-sections --variable theme=bootstrap --css /__w/_temp/Library/BiocStyle/resources/html/bioconductor.css --include-in-header /tmp/Rtmpnpd80O/rmarkdown-str83f6c578a64.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --variable code_folding=hide --variable code_menu=1 
#> 
#> Output created: basicExploration.html

if (interactive()) {
    ## Browse the report
    browseURL(report)
}

## See the example output at
## http://leekgroup.github.io/regionReport/reference/derfinderReport-example/basicExploration/basicExploration.html
if (FALSE) {
## Note that you can run the example using:
example("derfinderReport", "regionReport", ask = FALSE)
}